requirements.txt
streamlit
pypdf
python-docx
groq
assemblyai
pandas
networkx
python-dotenv
matplotlib